Made suggested changes to exercise 2
