+++ exercises/joseph_ramsay/ex02_oo_basics/src/Shape.cpp
@@ -1,0 +1,3 @@
+#include "Shape.h"
+
+Shape::~Shape() {}
