--- exercises/joseph_ramsay/ex02_oo_basics/src/shape.cpp
@@ -1,3 +1,0 @@
-#include "shape.h"
-
-int Shape::getNumSides() { return m_numSides; }
