--- exercises/joseph_ramsay/ex02_oo_basics/src/shapes.h
@@ -1,5 +1,0 @@
-#include "circle.h"
-#include "rectangle.h"
-#include "shape.h"
-#include "square.h"
-#include "triangle.h"
