--- exercises/joseph_ramsay/ex02_oo_basics/src/square.cpp
@@ -1,9 +1,0 @@
-#include "square.h"
-
-Square::Square(int sideLength) : Rectangle(sideLength, sideLength){};
-
-ShapeType Square::getType() const { return T_square; }
-
-std::string Square::toString() const {
-  return std::string("square(side=") + std::to_string(m_side1Length) + ")";
-}
